Break nccl_ofi_rdma.h into multiple headers
In order to make the RDMA structures easier to maintain and easier to refactor
into C++ classes, this breaks the nccl_ofi_rdma.h header into separate headers
for each transport component (endpoint, communicator, device, etc) and other
categories of components. Future commits can breakdown the headers more
granularly as needed.
