+++ src/nccl_ofi_rdma.cpp
@@ -21,7 +21,15 @@
 #endif
 #include "nccl_ofi_ep_addr_list.h"
 #include "nccl_ofi_param.h"
-#include "nccl_ofi_rdma.h"
+#include "rdma/nccl_ofi_rdma_communicator.h"
+#include "rdma/nccl_ofi_rdma_constants.h"
+#include "rdma/nccl_ofi_rdma_domain.h"
+#include "rdma/nccl_ofi_rdma_device.h"
+#include "rdma/nccl_ofi_rdma_endpoint.h"
+#include "rdma/nccl_ofi_rdma_messages.h"
+#include "rdma/nccl_ofi_rdma_mr_handle.h"
+#include "rdma/nccl_ofi_rdma_plugin.h"
+#include "rdma/nccl_ofi_rdma_request.h"
 #include "nccl_ofi_math.h"
 #include "nccl_ofi_tracepoint.h"
 #include "nccl_ofi_scheduler.h"
