+++ src/platform-aws.cpp
@@ -24,7 +24,8 @@
 #include "platform-aws.h"
 #include "nccl_ofi_log.h"
 #include "nccl_ofi_math.h"
-#include "nccl_ofi_rdma.h"
+#include "rdma/nccl_ofi_rdma_communicator.h"
+#include "rdma/nccl_ofi_rdma_messages.h"
 #include "nccl_ofi_param.h"
 #include "nccl_ofi_pthread.h"
 #include "nccl_ofi_system.h"
